Add AV.Decoder#decode method to wrap around the readChunk hook.
Decoders now no longer have to emit data/error events from readChunk, they just throw errors and return the result, and decode handles the rest. If an AV.UnderflowError is thrown, the packet is automatically retried when more data is available. Other errors are emitted as 'error' events. If the end of decoding is reached, an 'end' event is emitted.