Make Demuxer#readChunk behave more like Decoder#decodePacket
Many media formats are structured in a series of discrete chunks. It makes sense for the Demuxer#readChunk implementation to read a single chunk, rather than having to deal with the complexities of streamed data. Demuxer#readChunk will be called as necessary to consume data in the stream. If an underflow occurs, it will seek back to the last good offset, and try again when more data is available. This is the same behavior that decoders already have.